requirements.txt
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cmake==3.27.0
decorator==4.4.2
filelock==3.12.2
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
install==1.3.5
Jinja2==3.1.2
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.0.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper @ git+https://github.com/openai/whisper.git@b91c907694f96a3fb9da03d4bbdc83fbcd3a40a4
Pillow==10.0.0
proglog==0.1.10
pysrt==1.1.2
pyTelegramBotAPI==4.12.0
pytube==15.0.0
regex==2023.6.3
requests==2.31.0
sympy==1.12
telebot==0.0.5
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
urllib3==2.0.4
